telegram_bot_backend/app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.2.0"
description = ""
authors = ["waffle <b07902011@ntu.edu.tw>"]
packages = [{ include = "**/*.py" }]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.11"
fastapi = "^0.75.0"
uvicorn = "^0.17.5"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
